Make JsonContentOfT derive directly from HttpContent
