+++ src/libraries/System.Net.Http.Json/ref/System.Net.Http.Json.cs
@@ -40,7 +40,7 @@ namespace System.Net.Http.Json
         public static System.Threading.Tasks.Task<T?> ReadFromJsonAsync<T>(this System.Net.Http.HttpContent content, System.Text.Json.JsonSerializerOptions? options = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
         public static System.Threading.Tasks.Task<T?> ReadFromJsonAsync<T>(this System.Net.Http.HttpContent content, System.Text.Json.Serialization.Metadata.JsonTypeInfo<T> jsonTypeInfo, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
     }
-    public partial class JsonContent : System.Net.Http.HttpContent
+    public sealed partial class JsonContent : System.Net.Http.HttpContent
     {
         internal JsonContent() { }
         public System.Type ObjectType { get { throw null; } }
