+++ src/libraries/System.Net.Http.Json/src/System/Net/Http/Json/HttpContentJsonExtensions.cs
@@ -20,7 +20,7 @@ namespace System.Net.Http.Json
                 throw new ArgumentNullException(nameof(content));
             }
 
-            Encoding? sourceEncoding = JsonContent.GetEncoding(content.Headers.ContentType?.CharSet);
+            Encoding? sourceEncoding = Helper.GetEncoding(content.Headers.ContentType?.CharSet);
 
             return ReadFromJsonAsyncCore(content, type, sourceEncoding, options, cancellationToken);
         }
@@ -32,7 +32,7 @@ namespace System.Net.Http.Json
                 throw new ArgumentNullException(nameof(content));
             }
 
-            Encoding? sourceEncoding = JsonContent.GetEncoding(content.Headers.ContentType?.CharSet);
+            Encoding? sourceEncoding = Helper.GetEncoding(content.Headers.ContentType?.CharSet);
 
             return ReadFromJsonAsyncCore<T>(content, sourceEncoding, options, cancellationToken);
         }
@@ -60,7 +60,7 @@ namespace System.Net.Http.Json
                 throw new ArgumentNullException(nameof(content));
             }
 
-            Encoding? sourceEncoding = JsonContent.GetEncoding(content.Headers.ContentType?.CharSet);
+            Encoding? sourceEncoding = Helper.GetEncoding(content.Headers.ContentType?.CharSet);
 
             return ReadFromJsonAsyncCore(content, type, sourceEncoding, context, cancellationToken);
         }
@@ -72,7 +72,7 @@ namespace System.Net.Http.Json
                 throw new ArgumentNullException(nameof(content));
             }
 
-            Encoding? sourceEncoding = JsonContent.GetEncoding(content.Headers.ContentType?.CharSet);
+            Encoding? sourceEncoding = Helper.GetEncoding(content.Headers.ContentType?.CharSet);
 
             return ReadFromJsonAsyncCore<T>(content, sourceEncoding, jsonTypeInfo, cancellationToken);
         }
