+++ src/libraries/System.Net.Http.Json/src/System/Net/Http/Json/JsonContent.cs
@@ -1,7 +1,9 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+#if !NETCOREAPP
 using System.Diagnostics;
+#endif
 using System.IO;
 using System.Net.Http.Headers;
 using System.Text;
@@ -11,19 +13,15 @@ using System.Threading.Tasks;
 
 namespace System.Net.Http.Json
 {
-    public partial class JsonContent : HttpContent
+    public sealed partial class JsonContent : HttpContent
     {
-        private JsonSerializerOptions _jsonSerializerOptions = null!;
+        internal static readonly JsonSerializerOptions s_defaultSerializerOptions = new JsonSerializerOptions(JsonSerializerDefaults.Web);
 
-        private static MediaTypeHeaderValue DefaultMediaType => new MediaTypeHeaderValue("application/json") { CharSet = "utf-8" };
-
-        internal static JsonSerializerOptions s_defaultSerializerOptions { get; } = new JsonSerializerOptions(JsonSerializerDefaults.Web);
-
+        private readonly JsonSerializerOptions? _jsonSerializerOptions;
         public Type ObjectType { get; }
-
         public object? Value { get; }
 
-        internal JsonContent(object? inputValue, Type inputType, MediaTypeHeaderValue? mediaType)
+        private JsonContent(object? inputValue, Type inputType, MediaTypeHeaderValue? mediaType, JsonSerializerOptions? options)
         {
             if (inputType == null)
             {
@@ -37,18 +35,15 @@ namespace System.Net.Http.Json
 
             Value = inputValue;
             ObjectType = inputType;
-            Headers.ContentType = mediaType ?? DefaultMediaType;
+            Headers.ContentType = mediaType ?? Helper.GetDefaultMediaType();
+            _jsonSerializerOptions = options ?? s_defaultSerializerOptions;
         }
 
         public static JsonContent Create<T>(T inputValue, MediaTypeHeaderValue? mediaType = null, JsonSerializerOptions? options = null)
             => Create(inputValue, typeof(T), mediaType, options);
 
         public static JsonContent Create(object? inputValue, Type inputType, MediaTypeHeaderValue? mediaType = null, JsonSerializerOptions? options = null)
-        {
-            JsonContent content = new JsonContent(inputValue, inputType, mediaType);
-            content._jsonSerializerOptions = options ?? s_defaultSerializerOptions;
-            return content;
-        }
+            => new JsonContent(inputValue, inputType, mediaType, options);
 
         protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context)
             => SerializeToStreamAsyncCore(stream, async: true, CancellationToken.None);
@@ -61,9 +56,7 @@ namespace System.Net.Http.Json
 
         private async Task SerializeToStreamAsyncCore(Stream targetStream, bool async, CancellationToken cancellationToken)
         {
-            Debug.Assert(_jsonSerializerOptions != null);
-
-            Encoding? targetEncoding = GetEncoding(Headers.ContentType?.CharSet);
+            Encoding? targetEncoding = Helper.GetEncoding(Headers.ContentType?.CharSet);
 
             // Wrap provided stream into a transcoding stream that buffers the data transcoded from utf-8 to the targetEncoding.
             if (targetEncoding != null && targetEncoding != Encoding.UTF8)
@@ -128,35 +121,6 @@ namespace System.Net.Http.Json
 #endif
                 }
             }
-        }
-
-        internal static Encoding? GetEncoding(string? charset)
-        {
-            Encoding? encoding = null;
-
-            if (charset != null)
-            {
-                try
-                {
-                    // Remove at most a single set of quotes.
-                    if (charset.Length > 2 && charset[0] == '\"' && charset[charset.Length - 1] == '\"')
-                    {
-                        encoding = Encoding.GetEncoding(charset.Substring(1, charset.Length - 2));
-                    }
-                    else
-                    {
-                        encoding = Encoding.GetEncoding(charset);
-                    }
-                }
-                catch (ArgumentException e)
-                {
-                    throw new InvalidOperationException(SR.CharSetInvalid, e);
-                }
-
-                Debug.Assert(encoding != null);
-            }
-
-            return encoding;
         }
     }
 }
