+++ src/libraries/System.Net.Http.Json/src/System/Net/Http/Json/JsonContent.netcoreapp.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace System.Net.Http.Json
 {
-    public partial class JsonContent
+    public sealed partial class JsonContent
     {
         protected override void SerializeToStream(Stream stream, TransportContext? context, CancellationToken cancellationToken)
             => SerializeToStreamAsyncCore(stream, async: false, cancellationToken).GetAwaiter().GetResult();
