+++ src/libraries/System.Net.Http.Json/src/System/Net/Http/Json/JsonContentOfT.netcoreapp.cs
@@ -9,17 +9,9 @@ namespace System.Net.Http.Json
 {
     internal sealed partial class JsonContent<TValue>
     {
-        /// <summary>
-        /// This method is overriden to avoid accidentally rooting <see cref="JsonContent.SerializeToStreamAsyncCore(Stream, bool, CancellationToken)"/> post ILLinker trimming.
-        /// See <see cref="JsonContent{TValue}.SerializeToStreamAsyncCore(Stream, bool, CancellationToken)"/> for more info.
-        /// </summary>
         protected override void SerializeToStream(Stream stream, TransportContext? context, CancellationToken cancellationToken)
             => SerializeToStreamAsyncCore(stream, async: false, cancellationToken).GetAwaiter().GetResult();
 
-        /// <summary>
-        /// This method is overriden to avoid accidentally rooting <see cref="JsonContent.SerializeToStreamAsyncCore(Stream, bool, CancellationToken)"/> post ILLinker trimming.
-        /// See <see cref="JsonContent{TValue}.SerializeToStreamAsyncCore(Stream, bool, CancellationToken)"/> for more info.
-        /// </summary>
         protected override Task SerializeToStreamAsync(Stream stream, TransportContext? context, CancellationToken cancellationToken)
             => SerializeToStreamAsyncCore(stream, async: true, cancellationToken);
     }
